requirements.txt
numpy~=1.22.4
Pillow~=9.2.0
numba~=0.55.2